Compute quaternion to mat directly
Compute quaternion to mat using a direct formula to avoid an annoying
bug: when rotation is close to null, the quaternion would look like :
    (1., 0., 0., 0.)
Which would result in normalizing a null vector in load_rot(), and a
cancerous NaN propagation, essentially leading to a black screen when a
camera is set to look towards -z.

This method is also substantially faster and leverage the full power of
quaternions.