First basic animation engine
An animation targets a single node and is composed of tracks. A track is
a set of keyframes, a tuple of time and value, targeting a parameter in
the target node such as position on one axis, rotation on one axis or
scale on one axis.

Tracks can be either linear of Bézier curves.

A clip is a set of animations that are meant to be played together. A
clip can be played forward, backward or back amd forth (once forward
then backward). A Clip can be looped indefinitely.

The user has to allocate a slot in the animQueue. A slot is an
animStack, where the user can push or append Clips to be played at
certain times.