[ogex] slew camera only if up is AXIS_Z
A dirty hack currently slews the camera 90d down to compensate for some
weird blender export artefact. This is probably due to axis
interpretation: blender has an identity cam to look toward -z. Our
identity cam looks toward -y. When we swap Y and Z axis, we both modify
the vertex info and the transforms. However, an identity matrix remains
identity through this transformation. In other words, while we move the
whole word around the camera, it stays the same, and so, in our case, it
ends up looking toward -z when we expect it to look toward -y.

Another way to think of it is that the slew is the camera equivalent of
inverting y and z values of geometry objects.