[ogex] reset node's model when skinned
This fixes a bug where the transform of a skinned node would be applied
twice. Indeed, for some reasons opengex sets the Skin transform equals
to the skinned node transform, and requires the Skin transform to be
applied when computing skinned vertices positions, which would then be
multiplied by the node's transform, hence a redundant transformation.

Now when a node is skinned, it is just reset to an identity
transformation.