[ogex] split parsing in 3 phases (pre, -, post)
We now parse top-level structures such as Material, GeometryObject and
CameraObject before all other structures, and we added a post-parsing
phase to resolve skeletons (because they require the whole graph to be
parsed and valid). This makes the parser more robust to arbitrary
ordering of structures, while also making the whole process simpler.