refactor anim engine
It is now possible to animate pos, scale, rot and quat in a single
track. Animations do not contain a full trackset anymore but instead a
variable size array of tracks targetting any channel on demand.

The API was modified to be able to add tracks easily within an
Animation.

Those modifications make it much easier to import/create complex
animations, and make the animation engine overall faster and more
efficient while being simpler to use.